requirements.txt
requests
pyodbc
azure-functions
python-dotenv
sumy
nltk